Refactor legend generation into modular helper functions
* Extract legend processing logic into separate helper modules
* Add new helper files for category, equal number, and equal interval legend handling
* Create legendUtils for common legend initialization and finalization
* Add legendDataProcessing for data filtering and preparation
* Update MapConfig type to include roundToPlace property for primary columns
* Add comprehensive test file for generateRuntimeLegend
* Improve type safety with better TypeScript definitions
* Maintain backward compatibility while improving code organization
